Feature: Could not login with WalletConnect
Solution: When a user wants to connect with his wallet, a modal appears and offers him 2 ways to connect:

- Metamask
- Wallet Connect

If the user chooses metamask, the schema remains the same, the metamask popup appears and asks him to connect an account.
If Wallet Connect is selected, a QR Code dedicated to the connection appears.